[UHYU-178] refactor: (코드레빗) 순환참조 제거 및 type하고 공백
