+++ src/features/benefit/types/index.ts
@@ -1,3 +1,3 @@
-export type{ BrandDetail } from './BrandDetail.types';
+export type { BrandDetail } from './BrandDetail.types';
 export type { PaginationProps } from './Pagination.types';
 export type { PaginationButtonProps } from './PaginationButton.types';
