+++ src/shared/components/bottom_sheet/BaseBottomSheet.tsx
@@ -1,4 +1,5 @@
-import { IconButton } from '@/shared/components';
+// 순환 의존성 방지: 컴포넌트 단일 경로로 직접 import
+import { IconButton } from '@/shared/components/buttons/IconButton';
 import { AnimatePresence, motion } from 'framer-motion';
 import { ChevronLeft, X } from 'lucide-react';
 import React, { useEffect } from 'react';
