+++ src/shared/components/buttons/FileUploadButton.tsx
@@ -1,4 +1,4 @@
-import { ButtonBase } from '@/shared/components';
+import { ButtonBase } from '@/shared/components/buttons/ButtonBase';
 import { useRef } from 'react';
 
 interface FileUploadButtonProps {
